requirements.txt
fastapi
uvicorn
transformers
torch
torchaudio
moviepy
TTS
deep-translator
ffmpeg-python
gtts
boto3